feat!: add random_kinesis_event()
