+++ src/test/Cargo.toml
@@ -11,7 +11,6 @@ edition = "2018"
 arrow = { git = "https://github.com/DSLAM-UMD/arrow", branch = "scq", features = [ "simd" ] }
 base64 = "0.13.0"
 chrono = { version = "0.4.19", features = [ "serde" ] }
-
 datafusion = { git = "https://github.com/DSLAM-UMD/arrow", branch = "scq", features = [ "simd" ] }
 fake = { version = "2.4", features = [ 'derive', 'chrono' ] }
 rand = "0.8"
@@ -21,6 +20,8 @@ serde_bytes = "0.11"
 serde_json = "1.0"
 
 [dev-dependencies]
+aws_lambda_events = "0.4"
+
 tokio = { version = "1.2", features = [ "macros", "io-util", "sync", "rt-multi-thread" ] }
 
 [lib]
