+++ src/test/src/lib.rs
@@ -28,231 +28,21 @@
 )]
 #![feature(get_mut_unchecked)]
 
-use base64::{decode, encode};
-use chrono::{DateTime, TimeZone, Utc};
-use fake::{Dummy, Fake, Faker};
-use rand::Rng;
-use serde::de::{Deserializer, Error as DeError};
-use serde::ser::Serializer;
+use fake::{Dummy, Fake};
+use runtime::prelude::Result;
 use serde::{Deserialize, Serialize};
-use std::ops::{Deref, DerefMut};
+use serde_json::Value;
 
+mod kinesis;
+
+/// The data record schema for unit tests.
 #[derive(Dummy, Debug, Clone, PartialEq, Deserialize, Serialize)]
-struct KinesisEvent {
-    #[serde(rename = "Records")]
-    #[dummy(faker = "(Faker, 1000000)")]
-    pub records: Vec<KinesisEventRecord>,
-}
-
-#[derive(Dummy, Debug, Clone, PartialEq, Deserialize, Serialize)]
-struct KinesisEventRecord {
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "awsRegion")]
-    pub aws_region:          Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "eventID")]
-    pub event_id:            Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "eventName")]
-    pub event_name:          Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "eventSource")]
-    pub event_source:        Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "eventSourceARN")]
-    pub event_source_arn:    Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "eventVersion")]
-    pub event_version:       Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "invokeIdentityArn")]
-    pub invoke_identity_arn: Option<String>,
-    pub kinesis:             KinesisRecord,
-}
-
-#[derive(Dummy, Debug, Clone, PartialEq, Deserialize, Serialize)]
-struct KinesisRecord {
-    #[serde(rename = "approximateArrivalTimestamp")]
-    pub approximate_arrival_timestamp: SecondTimestamp,
-    pub data:                          Base64Data,
-    #[serde(rename = "encryptionType")]
-    pub encryption_type:               Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "partitionKey")]
-    pub partition_key:                 Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "sequenceNumber")]
-    pub sequence_number:               Option<String>,
-    #[serde(deserialize_with = "deserialize_lambda_string")]
-    #[serde(default)]
-    #[serde(rename = "kinesisSchemaVersion")]
-    pub kinesis_schema_version:        Option<String>,
-}
-
-/// Deserializes `Option<String>`, mapping JSON `null` or the empty string `""`
-/// to `None`.
-#[cfg(not(feature = "string-null-empty"))]
-fn deserialize_lambda_string<'de, D>(
-    deserializer: D,
-) -> core::result::Result<Option<String>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    match Option::deserialize(deserializer)? {
-        Some(s) => {
-            if s == "" {
-                Ok(None)
-            } else {
-                Ok(Some(s))
-            }
-        }
-        None => Ok(None),
-    }
-}
-
-/// Randomly dataset schema.
-#[derive(Dummy, Debug, Clone, PartialEq, Deserialize, Serialize)]
-struct DataRecord {
-    #[dummy(faker = "1..100000")]
+pub(crate) struct DataRecord {
+    #[dummy(faker = "1..100")]
     pub c1: i64,
-    #[dummy(faker = "1..50000")]
+    #[dummy(faker = "1..100")]
     pub c2: i64,
     pub c3: String,
-}
-
-/// Binary data encoded in base64.
-#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
-struct Base64Data(
-    #[serde(deserialize_with = "deserialize_base64")]
-    #[serde(serialize_with = "serialize_base64")]
-    pub Vec<u8>,
-);
-
-impl Dummy<Faker> for Base64Data {
-    fn dummy_with_rng<R: Rng + ?Sized>(_: &Faker, _: &mut R) -> Self {
-        let record: DataRecord = Faker.fake();
-        Base64Data(serde_json::to_string(&record).unwrap().into_bytes())
-    }
-}
-
-impl Deref for Base64Data {
-    type Target = Vec<u8>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for Base64Data {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-fn deserialize_base64<'de, D>(deserializer: D) -> core::result::Result<Vec<u8>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let s: String = String::deserialize(deserializer)?;
-    decode(&s).map_err(DeError::custom)
-}
-
-fn serialize_base64<S>(value: &[u8], serializer: S) -> core::result::Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    serializer.serialize_str(&encode(value))
-}
-
-/// Timestamp with second precision.
-#[derive(Dummy, Debug, Clone, PartialEq, Deserialize, Serialize)]
-struct SecondTimestamp(
-    #[serde(deserialize_with = "deserialize_seconds")]
-    #[serde(serialize_with = "serialize_seconds")]
-    pub DateTime<Utc>,
-);
-
-impl Deref for SecondTimestamp {
-    type Target = DateTime<Utc>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for SecondTimestamp {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-fn serialize_seconds<S>(
-    date: &DateTime<Utc>,
-    serializer: S,
-) -> core::result::Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    let seconds = date.timestamp();
-    let milliseconds = date.timestamp_subsec_millis();
-    let combined = format!("{}.{}", seconds, milliseconds);
-    if milliseconds > 0 {
-        serializer.serialize_str(&combined)
-    } else {
-        serializer.serialize_str(&seconds.to_string())
-    }
-}
-
-#[allow(dead_code)]
-fn deserialize_seconds<'de, D>(deserializer: D) -> core::result::Result<DateTime<Utc>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let (whole, frac) = normalize_timestamp(deserializer)?;
-    let seconds = whole;
-    let nanos = frac * 1_000_000;
-    Ok(Utc.timestamp(seconds as i64, nanos as u32))
-}
-
-fn normalize_timestamp<'de, D>(deserializer: D) -> core::result::Result<(u64, u64), D::Error>
-where
-    D: Deserializer<'de>,
-{
-    #[derive(Deserialize)]
-    #[serde(untagged)]
-    enum StringOrNumber {
-        String(String),
-        Float(f64),
-        Int(u64),
-    }
-
-    let input: f64 = match StringOrNumber::deserialize(deserializer)? {
-        StringOrNumber::String(s) => s.parse::<f64>().map_err(DeError::custom)?,
-        StringOrNumber::Float(f) => f,
-        StringOrNumber::Int(i) => i as f64,
-    };
-
-    // We need to do this due to floating point issues.
-    let input_as_string = format!("{}", input);
-    let parts: core::result::Result<Vec<u64>, _> = input_as_string
-        .split('.')
-        .map(|x| x.parse::<u64>().map_err(DeError::custom))
-        .collect();
-    let parts = parts?;
-    if parts.len() > 1 {
-        Ok((parts[0], parts[1]))
-    } else {
-        Ok((parts[0], 0))
-    }
 }
 
 /// Compares formatted output of a record batch with an expected
@@ -322,19 +112,84 @@ macro_rules! assert_batches_sorted_eq {
     };
 }
 
+/// Generate a random Kinesis event.
+///
+/// You can use an AWS Lambda function to process records in an Amazon Kinesis
+/// data stream. A Kinesis data stream is a set of shards. Each shard contains a
+/// sequence of data records. Before invoking the function, Lambda continues to
+/// read records from the stream until it has gathered a full batch, or until
+/// the batch window expires.
+///
+/// More details: <https://docs.aws.amazon.com/lambda/latest/dg/with-kinesis.html>
+///
+/// # Arguments
+///
+/// * `size`: the number of records in the event.
+///
+/// # Return
+///
+/// A valid JSON [value](serde_json::Value) representing the Kinesis event.
+///
+/// # Example
+///
+/// Kinesis record event
+///
+/// ```json
+/// {
+///     "Records": [
+///         {
+///             "kinesis": {
+///                 "kinesisSchemaVersion": "1.0",
+///                 "partitionKey": "1",
+///                 "sequenceNumber": "49590338271490256608559692538361571095921575989136588898",
+///                 "data": "SGVsbG8sIHRoaXMgaXMgYSB0ZXN0Lg==",
+///                 "approximateArrivalTimestamp": 1545084650.987
+///             },
+///             "eventSource": "aws:kinesis",
+///             "eventVersion": "1.0",
+///             "eventID": "shardId-000000000006:49590338271490256608559692538361571095921575989136588898",
+///             "eventName": "aws:kinesis:record",
+///             "invokeIdentityArn": "arn:aws:iam::123456789012:role/lambda-role",
+///             "awsRegion": "us-east-2",
+///             "eventSourceARN": "arn:aws:kinesis:us-east-2:123456789012:stream/lambda-stream"
+///         },
+///         {
+///             "kinesis": {
+///                 "kinesisSchemaVersion": "1.0",
+///                 "partitionKey": "1",
+///                 "sequenceNumber": "49590338271490256608559692540925702759324208523137515618",
+///                 "data": "VGhpcyBpcyBvbmx5IGEgdGVzdC4=",
+///                 "approximateArrivalTimestamp": 1545084711.166
+///             },
+///             "eventSource": "aws:kinesis",
+///             "eventVersion": "1.0",
+///             "eventID": "shardId-000000000006:49590338271490256608559692540925702759324208523137515618",
+///             "eventName": "aws:kinesis:record",
+///             "invokeIdentityArn": "arn:aws:iam::123456789012:role/lambda-role",
+///             "awsRegion": "us-east-2",
+///             "eventSourceARN": "arn:aws:kinesis:us-east-2:123456789012:stream/lambda-stream"
+///         }
+///     ]
+/// }
+/// ```
+pub fn random_kinesis_event(size: usize) -> Result<Value> {
+    Ok(serde_json::to_value(kinesis::KinesisEvent {
+        records: fake::vec![kinesis::KinesisEventRecord; size],
+    })?)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use runtime::prelude::*;
 
     #[tokio::test]
-    async fn fake_data() -> Result<()> {
-        let record: KinesisRecord = Faker.fake();
-        println!("{:?}", record);
-
-        let record = serde_json::to_string(&record)?;
-        println!("{}", record);
-
+    async fn random_kinesis_data() -> Result<()> {
+        for i in 1..5 {
+            let value = random_kinesis_event(i)?;
+            let event: aws_lambda_events::event::kinesis::KinesisEvent =
+                serde_json::from_value(value)?;
+            assert_eq!(event.records.len(), i);
+        }
         Ok(())
     }
 }
